// VIS_Application/app/Model/HumanResource/Attendance/EmployeeFeedback.ts
export interface IEmployeeFeedback {
    Id: number;
    UserID: number;
    Remarks: string;
    IsLike: string;
    CreatedOn: string;
    EmployeeName: string;
    SenderName: string;
    TeamEmployeeId: number;

}
export interface IEmployee {

    EmployeeId: number;
    Employee_Name: string;

}